requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
holidays>=0.53
pyyaml>=6.0
tqdm>=4.66
# Optional: PyTorch CPU only (install separately as below)
